requirements.txt
fastapi
jinja2
uvicorn
python-dotenv
boto3
pydantic
python-dotenv
requests
python-multipart
livekit-agents[openai,deepgram,silero,turn-detector]~=1.2
livekit-plugins-noise-cancellation~=0.2